util func for resourceGroup name; remove aad, aof and rdb Redis settings

+++ pkg/kcp/provider/azure/redisinstance/deleteResourceGroup.go
@@ -6,6 +6,7 @@ import (
 	"github.com/kyma-project/cloud-manager/api/cloud-control/v1beta1"
 	"github.com/kyma-project/cloud-manager/pkg/composed"
 	azuremeta "github.com/kyma-project/cloud-manager/pkg/kcp/provider/azure/meta"
+	azureUtil "github.com/kyma-project/cloud-manager/pkg/kcp/provider/azure/util"
 	"github.com/kyma-project/cloud-manager/pkg/util"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -19,9 +20,13 @@ func deleteResourceGroup(ctx context.Context, st composed.State) (error, context
 		return nil, nil
 	}
 
-	logger.Info("Deleting Azure Redis resourceGroupName")
+	if state.resourceGroup.State == "Deleting" {
+		return nil, nil
+	}
 
-	resourceGroupName := "cm.redis." + state.ObjAsRedisInstance().Name
+	logger.Info("Deleting Azure Redis resourceGroup")
+
+	resourceGroupName := azureUtil.GetResourceGroupName("redis", state.ObjAsRedisInstance().Name)
 
 	error := state.client.DeleteResourceGroup(ctx, resourceGroupName)
 	if error != nil {
